+++ components/collector/src/source_collectors/azure_devops.py
@@ -90,12 +90,12 @@ class AzureDevopsSourceUpToDateness(SourceCollector):
 
     def _api_url(self) -> URL:
         url = super()._api_url()
-        project = str(url).split("/")[-1]
+        repository = self._parameter("repository", quote=True)
         path = self._parameter("file_path", quote=True)
         branch = self._parameter("branch", quote=True)
         search_criteria = \
             f"searchCriteria.itemPath={path}&searchCriteria.itemVersion.version={branch}&searchCriteria.$top=1"
-        return URL(f"{url}/_apis/git/repositories/{project}/commits?{search_criteria}&api-version=4.1")
+        return URL(f"{url}/_apis/git/repositories/{repository}/commits?{search_criteria}&api-version=4.1")
 
     def _parse_source_responses_value(self, responses: Responses) -> Value:
         return str(days_ago(parse(responses[0].json()["value"][0]["committer"]["date"])))